requirements.txt
appnope==0.1.2; sys_platform == "darwin" and python_version >= "3.7" and platform_system == "Darwin"
argcomplete==1.12.3; python_version < "3.8.0" and python_version >= "3.7"
argon2-cffi==21.1.0; python_version >= "3.6"
asgiref==3.4.1; python_version >= "3.6"
attrs==21.2.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
backcall==0.2.0; python_version >= "3.7"
bleach==4.1.0; python_version >= "3.7"
cffi==1.14.6; implementation_name == "pypy" and python_version >= "3.6"
colorama==0.4.4; python_version >= "3.7" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.7" and python_full_version >= "3.5.0"
crispy-bootstrap5==0.5; python_version >= "3.6"
cython==0.29.22; python_version >= "3.4" and python_full_version < "3.0.0" and python_version < "3.10" or python_version >= "3.4" and python_version < "3.10" and python_full_version >= "3.3.0"
debugpy==1.4.1; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
decorator==5.0.9; python_version >= "3.7"
defusedxml==0.7.1; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
django-crispy-forms==1.12.0; python_version >= "3.6"
django-easy-select2==1.5.7
django-extensions==3.1.3; python_version >= "3.6"
django==3.2.7; python_version >= "3.6"
entrypoints==0.3; python_full_version >= "3.6.1" and python_version >= "3.7"
gunicorn==20.1.0; python_version >= "3.5"
importlib-metadata==4.8.1; python_version == "3.7"
ipykernel==6.4.0; python_version >= "3.7"
ipython-genutils==0.2.0; python_version >= "3.7"
ipython==7.27.0; python_version >= "3.7"
jedi==0.18.0; python_version >= "3.7"
jinja2==3.0.1; python_version >= "3.7"
jsonschema==3.2.0; python_version >= "3.6"
jupyter-client==7.0.2; python_full_version >= "3.6.1" and python_version >= "3.7"
jupyter-core==4.7.1; python_full_version >= "3.6.1" and python_version >= "3.7"
jupyterlab-pygments==0.1.2; python_version >= "3.7"
markupsafe==2.0.1; python_version >= "3.6"
matplotlib-inline==0.1.3; python_version >= "3.7"
mazelib==0.9.12; python_version >= "3.4" and python_version < "3.10"
mistune==0.8.4; python_version >= "3.7"
nbclient==0.5.4; python_full_version >= "3.6.1" and python_version >= "3.7"
nbconvert==6.1.0; python_version >= "3.7"
nbformat==5.1.3; python_full_version >= "3.6.1" and python_version >= "3.7"
nest-asyncio==1.5.1; python_full_version >= "3.6.1" and python_version >= "3.7"
notebook==6.4.3; python_version >= "3.6"
numpy==1.20.1; python_version >= "3.7" and python_version < "3.10"
packaging==21.0; python_version >= "3.7"
pandocfilters==1.4.3; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.7"
parso==0.8.2; python_version >= "3.7"
pexpect==4.8.0; sys_platform != "win32" and python_version >= "3.7"
pickleshare==0.7.5; python_version >= "3.7"
prometheus-client==0.11.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
prompt-toolkit==3.0.20; python_full_version >= "3.6.2" and python_version >= "3.7"
psycopg2-binary==2.9.1; python_version >= "3.6"
ptyprocess==0.7.0; sys_platform != "win32" and python_version >= "3.7" and os_name != "nt"
py==1.10.0; python_version >= "3.6" and python_full_version < "3.0.0" and implementation_name == "pypy" or implementation_name == "pypy" and python_version >= "3.6" and python_full_version >= "3.4.0"
pycparser==2.20; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
pygments==2.10.0; python_version >= "3.7"
pyparsing==2.4.7; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.7"
pyrsistent==0.18.0; python_version >= "3.6"
python-dateutil==2.8.2; python_full_version >= "3.6.1" and python_version >= "3.6"
pytz==2021.1; python_version >= "3.6"
pywin32==301; sys_platform == "win32" and python_version >= "3.6"
pywinpty==1.1.4; os_name == "nt" and python_version >= "3.6"
pyzmq==22.2.1; python_full_version >= "3.6.1" and python_version >= "3.6"
send2trash==1.8.0; python_version >= "3.6"
six==1.16.0; python_full_version >= "3.6.1" and python_version >= "3.6" and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.7") and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6")
sqlparse==0.4.1; python_version >= "3.6"
terminado==0.12.1; python_version >= "3.6"
testpath==0.5.0; python_version >= "3.7"
tornado==6.1; python_full_version >= "3.6.1" and python_version >= "3.7"
traitlets==5.1.0; python_full_version >= "3.6.1" and python_version >= "3.7"
typing-extensions==3.10.0.2; python_version < "3.8" and python_version >= "3.7"
wcwidth==0.2.5; python_full_version >= "3.6.2" and python_version >= "3.7"
webencodings==0.5.1; python_version >= "3.7"
zipp==3.5.0; python_version < "3.8.0" and python_version >= "3.7"
